♻️ Rename the module to be more clear and meaningful.
