+++ test/unit_test/info_provider/mcp/test_protocol_runtime_checks.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from gearmeshing_ai.info_provider.mcp.client_async import AsyncMcpClient
+from gearmeshing_ai.info_provider.mcp.provider import AsyncMcpClient
 from gearmeshing_ai.info_provider.mcp.client_sync import McpClient
